Debita next attempt

+++ src/ProcessQueue.java
@@ -1,5 +1,5 @@
 
 public interface ProcessQueue {
 	public ProcessQueue addProcess(Process p);
-	public double countAverageWaitingTime();
+	public int countAverageWaitingTime();
 }
